errorcases: local/ runs against real docker stack; drop mocks & cluster/
Per request: the local-side error tests should exercise the live system,
not mocks. The docker-compose-based suite previously living in
errorcases/cluster/ becomes the local suite. The mock-based local tests
are removed — their assertions are subsumed by the integration tests.

- delete Mockito tests:
  * SubmittingExternalOrderErrorsLocalTest (the @WebMvcTest version)
  * FillOrderDispatcherErrorsLocalTest
- move cluster/ -> local/:
  * SubmittingExternalOrderErrorsClusterTest -> SubmittingExternalOrderErrorsLocalTest
  * ErrorsITSupport (package now ...errorcases.local)
- drop @EnabledIfSystemProperty(errorcases.cluster.it) — tests run on
  every `mvn test`; docker daemon is now a precondition for the suite
- remove the now-empty errorcases/cluster/ directory; per Tani's
  "ignore the cluster testing part" guidance we don't keep a stub here
